@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe "Posts", :type => :request do
+RSpec.describe "Posts", type: :request do
   describe "GET /posts" do
     it "works! (now write some real specs)" do
       get posts_path
