@@ -11,10 +11,19 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20160705002149) do
+ActiveRecord::Schema.define(version: 20160709150255) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
+
+  create_table "chat_messages", force: :cascade do |t|
+    t.text     "message"
+    t.integer  "user_id"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+  end
+
+  add_index "chat_messages", ["user_id"], name: "index_chat_messages_on_user_id", using: :btree
 
   create_table "emails", force: :cascade do |t|
     t.string   "subject"
@@ -126,6 +135,7 @@ ActiveRecord::Schema.define(version: 20160705002149) do
 
   add_index "users_roles", ["user_id", "role_id"], name: "index_users_roles_on_user_id_and_role_id", using: :btree
 
+  add_foreign_key "chat_messages", "users"
   add_foreign_key "emails", "users"
   add_foreign_key "identities", "users"
   add_foreign_key "posts", "users", column: "author_id"
