@@ -23,12 +23,16 @@ RSpec.describe PostsController, :type => :controller do
   # This should return the minimal set of attributes required to create a valid
   # Post. As you add validations to Post, be sure to
   # adjust the attributes here as well.
+  let (:user) {
+    FactoryGirl.create(:user)
+  }
+  
   let(:valid_attributes) {
-    skip("Add a hash of attributes valid for your model")
+    {title: 'Superpost', body: 'jshd dgb', author_id: user.id, is_pinned: true, is_draft: true, comments_count: 2}
   }
 
   let(:invalid_attributes) {
-    skip("Add a hash of attributes invalid for your model")
+    { title: nil, body: nil, author_id: user.id, is_pinned: true, is_draft: true, comments_count: 1 }
   }
 
   # This should return the minimal set of values that should be in the session
@@ -103,14 +107,14 @@ RSpec.describe PostsController, :type => :controller do
   describe "PUT update" do
     describe "with valid params" do
       let(:new_attributes) {
-        skip("Add a hash of attributes valid for your model")
+        {title: 'Superpost2', body: 'jshd dgb swf', author_id: user.id, is_pinned: true, is_draft: true, comments_count: 2}
       }
 
       it "updates the requested post" do
         post = Post.create! valid_attributes
         put :update, {:id => post.to_param, :post => new_attributes}, valid_session
         post.reload
-        skip("Add assertions for updated state")
+        expect(assigns(:post)).to eq(post)
       end
 
       it "assigns the requested post as @post" do
