@@ -1,4 +1,4 @@
 require 'rails_helper'
-RSpec.describe EmailWorker, type: :worker do
+RSpec.describe EmailWorkerWorker, type: :worker do
     pending "add some examples to (or delete) #{__FILE__}"
 end
